feat : spa로 게시물마다 상세보기 링크 생성
spa로 코드 자체는 board안에서 진행되지만 링크는 따로 부여
postid를 파라미터로 부여하는데 시간이 걸림
