+++ src/states/board/BoardStore.ts
@@ -1,0 +1,274 @@
+// 게시물 인터페이스 정의
+export interface Post {
+  boardId: number;
+  memberId: number;
+  category: string;
+  title: string; // 제목 추가
+  content: string;
+  createdAt: string;
+  likes: number;
+  likedByUser: boolean; // 게시물이 사용자가 좋아요를 눌렀는지 여부
+}
+
+// 게시물 데이터
+export const posts: Post[] = [
+  // IT 소식 카테고리
+  {
+    boardId: 1,
+    memberId: 1,
+    category: 'it-news',
+    title: 'IT 소식 제목 1',
+    content: 'IT 소식 내용 1',
+    createdAt: '2024-06-03T14:23:04',
+    likes: 10,
+    likedByUser: false,
+  },
+  {
+    boardId: 2,
+    memberId: 2,
+    category: 'it-news',
+    title: 'IT 소식 제목 2',
+    content: 'IT 소식 내용 2',
+    createdAt: '2024-06-03T14:24:05',
+    likes: 20,
+    likedByUser: false,
+  },
+  {
+    boardId: 3,
+    memberId: 3,
+    category: 'it-news',
+    title: 'IT 소식 제목 3',
+    content: 'IT 소식 내용 3',
+    createdAt: '2024-06-03T14:25:06',
+    likes: 5,
+    likedByUser: false,
+  },
+  {
+    boardId: 4,
+    memberId: 4,
+    category: 'it-news',
+    title: 'IT 소식 제목 4',
+    content: 'IT 소식 내용 4',
+    createdAt: '2024-06-03T14:26:07',
+    likes: 15,
+    likedByUser: false,
+  },
+  {
+    boardId: 5,
+    memberId: 5,
+    category: 'it-news',
+    title: 'IT 소식 제목 5',
+    content: 'IT 소식 내용 5',
+    createdAt: '2024-06-03T14:27:08',
+    likes: 8,
+    likedByUser: false,
+  },
+
+  // 잡담/일상 카테고리
+  {
+    boardId: 6,
+    memberId: 1,
+    category: 'chat',
+    title: '잡담 제목 1',
+    content: '잡담 내용 1',
+    createdAt: '2024-06-03T14:33:14',
+    likes: 10,
+    likedByUser: false,
+  },
+  {
+    boardId: 7,
+    memberId: 2,
+    category: 'chat',
+    title: '잡담 제목 2',
+    content: '잡담 내용 2',
+    createdAt: '2024-06-03T14:34:15',
+    likes: 20,
+    likedByUser: false,
+  },
+  {
+    boardId: 8,
+    memberId: 3,
+    category: 'chat',
+    title: '잡담 제목 3',
+    content: '잡담 내용 3',
+    createdAt: '2024-06-03T14:35:16',
+    likes: 5,
+    likedByUser: false,
+  },
+  {
+    boardId: 9,
+    memberId: 4,
+    category: 'chat',
+    title: '잡담 제목 4',
+    content: '잡담 내용 4',
+    createdAt: '2024-06-03T14:36:17',
+    likes: 15,
+    likedByUser: false,
+  },
+  {
+    boardId: 10,
+    memberId: 5,
+    category: 'chat',
+    title: '잡담 제목 5',
+    content: '잡담 내용 5',
+    createdAt: '2024-06-03T14:37:18',
+    likes: 8,
+    likedByUser: false,
+  },
+
+  // 기술 카테고리
+  {
+    boardId: 11,
+    memberId: 1,
+    category: 'tech',
+    title: '기술 제목 1',
+    content: '기술 내용 1',
+    createdAt: '2024-06-03T14:43:24',
+    likes: 10,
+    likedByUser: false,
+  },
+  {
+    boardId: 12,
+    memberId: 2,
+    category: 'tech',
+    title: '기술 제목 2',
+    content: '기술 내용 2',
+    createdAt: '2024-06-03T14:44:25',
+    likes: 20,
+    likedByUser: false,
+  },
+  {
+    boardId: 13,
+    memberId: 3,
+    category: 'tech',
+    title: '기술 제목 3',
+    content: '기술 내용 3',
+    createdAt: '2024-06-03T14:45:26',
+    likes: 5,
+    likedByUser: false,
+  },
+  {
+    boardId: 14,
+    memberId: 4,
+    category: 'tech',
+    title: '기술 제목 4',
+    content: '기술 내용 4',
+    createdAt: '2024-06-03T14:46:27',
+    likes: 15,
+    likedByUser: false,
+  },
+  {
+    boardId: 15,
+    memberId: 5,
+    category: 'tech',
+    title: '기술 제목 5',
+    content: '기술 내용 5',
+    createdAt: '2024-06-03T14:47:28',
+    likes: 8,
+    likedByUser: false,
+  },
+
+  // 인턴십/공모전 카테고리
+  {
+    boardId: 16,
+    memberId: 1,
+    category: 'internship',
+    title: '인턴십 제목 1',
+    content: '인턴십 내용 1',
+    createdAt: '2024-06-03T14:53:34',
+    likes: 10,
+    likedByUser: false,
+  },
+  {
+    boardId: 17,
+    memberId: 2,
+    category: 'internship',
+    title: '인턴십 제목 2',
+    content: '인턴십 내용 2',
+    createdAt: '2024-06-03T14:54:35',
+    likes: 20,
+    likedByUser: false,
+  },
+  {
+    boardId: 18,
+    memberId: 3,
+    category: 'internship',
+    title: '인턴십 제목 3',
+    content: '인턴십 내용 3',
+    createdAt: '2024-06-03T14:55:36',
+    likes: 5,
+    likedByUser: false,
+  },
+  {
+    boardId: 19,
+    memberId: 4,
+    category: 'internship',
+    title: '인턴십 제목 4',
+    content: '인턴십 내용 4',
+    createdAt: '2024-06-03T14:56:37',
+    likes: 15,
+    likedByUser: false,
+  },
+  {
+    boardId: 20,
+    memberId: 5,
+    category: 'internship',
+    title: '인턴십 제목 5',
+    content: '인턴십 내용 5',
+    createdAt: '2024-06-03T14:57:38',
+    likes: 8,
+    likedByUser: false,
+  },
+
+  // 공지사항 카테고리
+  {
+    boardId: 21,
+    memberId: 1,
+    category: 'notice',
+    title: '공지사항 제목 1',
+    content: '공지사항 내용 1',
+    createdAt: '2024-06-03T15:03:44',
+    likes: 10,
+    likedByUser: false,
+  },
+  {
+    boardId: 22,
+    memberId: 2,
+    category: 'notice',
+    title: '공지사항 제목 2',
+    content: '공지사항 내용 2',
+    createdAt: '2024-06-03T15:04:45',
+    likes: 20,
+    likedByUser: false,
+  },
+  {
+    boardId: 23,
+    memberId: 3,
+    category: 'notice',
+    title: '공지사항 제목 3',
+    content: '공지사항 내용 3',
+    createdAt: '2024-06-03T15:05:46',
+    likes: 5,
+    likedByUser: false,
+  },
+  {
+    boardId: 24,
+    memberId: 4,
+    category: 'notice',
+    title: '공지사항 제목 4',
+    content: '공지사항 내용 4',
+    createdAt: '2024-06-03T15:06:47',
+    likes: 15,
+    likedByUser: false,
+  },
+  {
+    boardId: 25,
+    memberId: 5,
+    category: 'notice',
+    title: '공지사항 제목 5',
+    content: '공지사항 내용 5',
+    createdAt: '2024-06-03T15:07:48',
+    likes: 8,
+    likedByUser: false,
+  },
+];
