--- src/states/board/BoardStore.tsx
@@ -1,524 +1,0 @@
-// 게시물 인터페이스 정의
-export interface Post {
-  boardId: number;
-  memberId: number;
-  category: string;
-  content: string;
-  createdAt: string;
-  likes: number;
-  likedByUser: boolean; // 게시물이 사용자가 좋아요를 눌렀는지 여부
-  myPost: boolean; // 사용자가 작성한 게시물인지 여부
-}
-
-// 게시물 데이터
-export const posts: Post[] = [
-  // IT 소식 카테고리
-  {
-    boardId: 1,
-    memberId: 1,
-    category: 'it-news',
-    content: 'IT 소식 내용 1',
-    createdAt: '2024-06-03T14:23:04',
-    likes: 10,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 2,
-    memberId: 2,
-    category: 'it-news',
-    content: 'IT 소식 내용 2',
-    createdAt: '2024-06-03T14:24:05',
-    likes: 20,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 3,
-    memberId: 3,
-    category: 'it-news',
-    content: 'IT 소식 내용 3',
-    createdAt: '2024-06-03T14:25:06',
-    likes: 5,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 4,
-    memberId: 4,
-    category: 'it-news',
-    content: 'IT 소식 내용 4',
-    createdAt: '2024-06-03T14:26:07',
-    likes: 15,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 5,
-    memberId: 5,
-    category: 'it-news',
-    content: 'IT 소식 내용 5',
-    createdAt: '2024-06-03T14:27:08',
-    likes: 8,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 6,
-    memberId: 6,
-    category: 'it-news',
-    content: 'IT 소식 내용 6',
-    createdAt: '2024-06-03T14:28:09',
-    likes: 18,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 7,
-    memberId: 7,
-    category: 'it-news',
-    content: 'IT 소식 내용 7',
-    createdAt: '2024-06-03T14:29:10',
-    likes: 3,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 8,
-    memberId: 8,
-    category: 'it-news',
-    content: 'IT 소식 내용 8',
-    createdAt: '2024-06-03T14:30:11',
-    likes: 13,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 9,
-    memberId: 9,
-    category: 'it-news',
-    content: 'IT 소식 내용 9',
-    createdAt: '2024-06-03T14:31:12',
-    likes: 7,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 10,
-    memberId: 10,
-    category: 'it-news',
-    content: 'IT 소식 내용 10',
-    createdAt: '2024-06-03T14:32:13',
-    likes: 17,
-    likedByUser: false,
-    myPost: false,
-  },
-
-  // 잡담/일상 카테고리
-  {
-    boardId: 11,
-    memberId: 1,
-    category: 'chat',
-    content: '잡담 내용 1',
-    createdAt: '2024-06-03T14:33:14',
-    likes: 10,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 12,
-    memberId: 2,
-    category: 'chat',
-    content: '잡담 내용 2',
-    createdAt: '2024-06-03T14:34:15',
-    likes: 20,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 13,
-    memberId: 3,
-    category: 'chat',
-    content: '잡담 내용 3',
-    createdAt: '2024-06-03T14:35:16',
-    likes: 5,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 14,
-    memberId: 4,
-    category: 'chat',
-    content: '잡담 내용 4',
-    createdAt: '2024-06-03T14:36:17',
-    likes: 15,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 15,
-    memberId: 5,
-    category: 'chat',
-    content: '잡담 내용 5',
-    createdAt: '2024-06-03T14:37:18',
-    likes: 8,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 16,
-    memberId: 6,
-    category: 'chat',
-    content: '잡담 내용 6',
-    createdAt: '2024-06-03T14:38:19',
-    likes: 18,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 17,
-    memberId: 7,
-    category: 'chat',
-    content: '잡담 내용 7',
-    createdAt: '2024-06-03T14:39:20',
-    likes: 3,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 18,
-    memberId: 8,
-    category: 'chat',
-    content: '잡담 내용 8',
-    createdAt: '2024-06-03T14:40:21',
-    likes: 13,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 19,
-    memberId: 9,
-    category: 'chat',
-    content: '잡담 내용 9',
-    createdAt: '2024-06-03T14:41:22',
-    likes: 7,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 20,
-    memberId: 10,
-    category: 'chat',
-    content: '잡담 내용 10',
-    createdAt: '2024-06-03T14:42:23',
-    likes: 17,
-    likedByUser: false,
-    myPost: false,
-  },
-
-  // 기술 카테고리
-  {
-    boardId: 21,
-    memberId: 1,
-    category: 'tech',
-    content: '기술 내용 1',
-    createdAt: '2024-06-03T14:43:24',
-    likes: 10,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 22,
-    memberId: 2,
-    category: 'tech',
-    content: '기술 내용 2',
-    createdAt: '2024-06-03T14:44:25',
-    likes: 20,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 23,
-    memberId: 3,
-    category: 'tech',
-    content: '기술 내용 3',
-    createdAt: '2024-06-03T14:45:26',
-    likes: 5,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 24,
-    memberId: 4,
-    category: 'tech',
-    content: '기술 내용 4',
-    createdAt: '2024-06-03T14:46:27',
-    likes: 15,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 25,
-    memberId: 5,
-    category: 'tech',
-    content: '기술 내용 5',
-    createdAt: '2024-06-03T14:47:28',
-    likes: 8,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 26,
-    memberId: 6,
-    category: 'tech',
-    content: '기술 내용 6',
-    createdAt: '2024-06-03T14:48:29',
-    likes: 18,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 27,
-    memberId: 7,
-    category: 'tech',
-    content: '기술 내용 7',
-    createdAt: '2024-06-03T14:49:30',
-    likes: 3,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 28,
-    memberId: 8,
-    category: 'tech',
-    content: '기술 내용 8',
-    createdAt: '2024-06-03T14:50:31',
-    likes: 13,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 29,
-    memberId: 9,
-    category: 'tech',
-    content: '기술 내용 9',
-    createdAt: '2024-06-03T14:51:32',
-    likes: 7,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 30,
-    memberId: 10,
-    category: 'tech',
-    content: '기술 내용 10',
-    createdAt: '2024-06-03T14:52:33',
-    likes: 17,
-    likedByUser: false,
-    myPost: false,
-  },
-
-  // 인턴십/공모전 카테고리
-  {
-    boardId: 31,
-    memberId: 1,
-    category: 'internship',
-    content: '인턴십 내용 1',
-    createdAt: '2024-06-03T14:53:34',
-    likes: 10,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 32,
-    memberId: 2,
-    category: 'internship',
-    content: '인턴십 내용 2',
-    createdAt: '2024-06-03T14:54:35',
-    likes: 20,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 33,
-    memberId: 3,
-    category: 'internship',
-    content: '인턴십 내용 3',
-    createdAt: '2024-06-03T14:55:36',
-    likes: 5,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 34,
-    memberId: 4,
-    category: 'internship',
-    content: '인턴십 내용 4',
-    createdAt: '2024-06-03T14:56:37',
-    likes: 15,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 35,
-    memberId: 5,
-    category: 'internship',
-    content: '인턴십 내용 5',
-    createdAt: '2024-06-03T14:57:38',
-    likes: 8,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 36,
-    memberId: 6,
-    category: 'internship',
-    content: '인턴십 내용 6',
-    createdAt: '2024-06-03T14:58:39',
-    likes: 18,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 37,
-    memberId: 7,
-    category: 'internship',
-    content: '인턴십 내용 7',
-    createdAt: '2024-06-03T14:59:40',
-    likes: 3,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 38,
-    memberId: 8,
-    category: 'internship',
-    content: '인턴십 내용 8',
-    createdAt: '2024-06-03T15:00:41',
-    likes: 13,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 39,
-    memberId: 9,
-    category: 'internship',
-    content: '인턴십 내용 9',
-    createdAt: '2024-06-03T15:01:42',
-    likes: 7,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 40,
-    memberId: 10,
-    category: 'internship',
-    content: '인턴십 내용 10',
-    createdAt: '2024-06-03T15:02:43',
-    likes: 17,
-    likedByUser: false,
-    myPost: false,
-  },
-
-  // 공지사항 카테고리
-  {
-    boardId: 41,
-    memberId: 1,
-    category: 'notice',
-    content: '공지사항 내용 1',
-    createdAt: '2024-06-03T15:03:44',
-    likes: 10,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 42,
-    memberId: 2,
-    category: 'notice',
-    content: '공지사항 내용 2',
-    createdAt: '2024-06-03T15:04:45',
-    likes: 20,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 43,
-    memberId: 3,
-    category: 'notice',
-    content: '공지사항 내용 3',
-    createdAt: '2024-06-03T15:05:46',
-    likes: 5,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 44,
-    memberId: 4,
-    category: 'notice',
-    content: '공지사항 내용 4',
-    createdAt: '2024-06-03T15:06:47',
-    likes: 15,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 45,
-    memberId: 5,
-    category: 'notice',
-    content: '공지사항 내용 5',
-    createdAt: '2024-06-03T15:07:48',
-    likes: 8,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 46,
-    memberId: 6,
-    category: 'notice',
-    content: '공지사항 내용 6',
-    createdAt: '2024-06-03T15:08:49',
-    likes: 18,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 47,
-    memberId: 7,
-    category: 'notice',
-    content: '공지사항 내용 7',
-    createdAt: '2024-06-03T15:09:50',
-    likes: 3,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 48,
-    memberId: 8,
-    category: 'notice',
-    content: '공지사항 내용 8',
-    createdAt: '2024-06-03T15:10:51',
-    likes: 13,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 49,
-    memberId: 9,
-    category: 'notice',
-    content: '공지사항 내용 9',
-    createdAt: '2024-06-03T15:11:52',
-    likes: 7,
-    likedByUser: false,
-    myPost: false,
-  },
-  {
-    boardId: 50,
-    memberId: 10,
-    category: 'notice',
-    content: '공지사항 내용 10',
-    createdAt: '2024-06-03T15:12:53',
-    likes: 17,
-    likedByUser: false,
-    myPost: false,
-  },
-];
